# docs/api/openapi/linksee-v1.yaml
openapi: 3.0.3
info:
  title: Linksee Teaching Collaboration API
  version: 1.0.0
  description: >
    Linksee 教学项目协作与过程评价平台的 P0 API schema。
    本文件与 docs/api/contracts/communication-contract-v1.md 保持一致，
    优先覆盖课程项目、阶段要求、小组讨论、阶段提交、教师反馈和课程看板。
servers:
  - url: /api/v1
    description: Linksee API v1
...
paths:
...
  /auth/change-password:
    post:
      tags:
        - Auth
      summary: 修改密码
      operationId: changePassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [currentPassword, newPassword]
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
                  minLength: 8
                  maxLength: 72
                  description: 必须包含大写字母、小写字母、数字
            example:
              currentPassword: "OldPass1"
              newPassword: "NewPass1"
      responses:
        "200":
          description: 密码修改成功，同时签发新 token
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: object
                    properties:
                      accessToken:
                        type: string
                      refreshToken:
                        type: string
                      expiresIn:
                        type: string
              example:
                ok: true
                data:
                  accessToken: "eyJ..."
                  refreshToken: "abc123"
                  expiresIn: "30m"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationFailed"
...
  /auth/admin/reset-password:
    post:
      tags:
        - Auth
      summary: 管理员或老师重置单个用户密码
      operationId: adminResetPassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [targetUserId]
              properties:
                targetUserId:
                  type: string
                newPassword:
                  type: string
                  description: 不传则系统随机生成
            example:
              targetUserId: "2024001001"
      responses:
        "200":
          description: 重置成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: object
                    properties:
                      temporaryPassword:
                        type: string
                      forceChangePassword:
                        type: boolean
              example:
                ok: true
                data:
                  temporaryPassword: "Ax8mPq29"
                  forceChangePassword: true
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /auth/admin/batch-reset-password:
    post:
      tags:
        - Auth
      summary: 教务按条件批量重置密码
      operationId: adminBatchResetPassword
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                newPassword:
                  type: string
                  description: 不传则系统随机生成
                userIds:
                  type: array
                  items:
                    type: string
                grade:
                  type: integer
                cohort:
                  type: integer
                adminClass:
                  type: string
                major:
                  type: string
            example:
              grade: 2023
              adminClass: 软工2301
      responses:
        "200":
          description: 批量重置成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  data:
                    type: object
                    properties:
                      defaultPassword:
                        type: string
                        nullable: true
                      affectedCount:
                        type: integer
                      forceChangePassword:
                        type: boolean
              example:
                ok: true
                data:
                  defaultPassword: "Nx3Qm27A"
                  affectedCount: 56
                  forceChangePassword: true
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthenticated"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthenticated:
      description: 未认证或 token 无效
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: UNAUTHENTICATED
            message: Authentication token is missing or invalid.
            details: null
            requestId: trace_20260430_001
    Forbidden:
      description: 当前用户无权访问该资源或执行该操作
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: FORBIDDEN
            message: Current user is not allowed to perform this operation.
            details: null
            requestId: trace_20260430_002
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: NOT_FOUND
            message: Requested resource was not found.
            details: null
            requestId: trace_20260430_003
...
    ValidationFailed:
      description: 请求字段校验失败
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            ok: false
            code: VALIDATION_FAILED
            message: Request payload validation failed.
            details: null
            requestId: trace_20260430_005
  schemas:
    ErrorResponse:
      type: object
      required:
        - ok
        - code
        - message
      properties:
        ok:
          type: boolean
          enum: [false]
        code:
          type: string
          enum:
            - UNAUTHENTICATED
            - FORBIDDEN
            - NOT_FOUND
            - CONFLICT
            - VALIDATION_FAILED
            - INTERNAL_ERROR
            - FORCE_CHANGE_PASSWORD
        message:
          type: string
        details:
          nullable: true
        requestId:
          type: string
          nullable: true
...